make `dlsym()` accept a byte length and callback
So that the user can request an explicit byte length of the
data the Buffer is representing. Previously it was fixed to
`sizeof(void *)`, which was just plain incorrect. Consider
an exported `char` which would only require 1 byte, or a
struct of some kind which would require more data than pointer
size.

Also now the user can pass a callback function in, which
will get invoked when the Buffer instance gets garbage
collected. I'm not sure of the exact use-case for this yet,
but I figured it could be useful for someone someday (I'm
thinking NodObjC with all of its weird cyclic refereces stuff).

What I'm not sure of yet is what arguments need to be passed
to the callback function, so for now nothing does. One option
would be to create *another* Buffer instance pointing to the
`data`, such that the user could operate on the pointer in
JS-land one final time for cleanup (or keep it for re-use),
but the garbage collection callback wouldn't be invoked in that
case. Still pondering that one…